+++ packages/protoplugin-test/src/transpile.test.ts
@@ -12,8 +12,31 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import { FileDescriptorProto } from "@bufbuild/protobuf";
-import { transpile } from "./helpers";
+import { CodeGeneratorRequest, FileDescriptorProto } from "@bufbuild/protobuf";
+import { createEcmaScriptPlugin } from "@bufbuild/protoplugin";
+import type { Schema } from "@bufbuild/protoplugin/ecmascript";
+
+export function transpile(
+  genTs: (schema: Schema) => void
+): (name: string) => string[] {
+  const req = new CodeGeneratorRequest({
+    parameter: `target=ts+js+dts`,
+  });
+  const plugin = createEcmaScriptPlugin({
+    name: "test-plugin",
+    version: "v99.0.0",
+    generateTs: genTs,
+  });
+  const res = plugin.run(req);
+  return function linesOf(filename: string): string[] {
+    const file = res.file.find((f) => f.name === filename);
+    if (!file) {
+      throw new Error(`did not find file ${filename}`);
+    }
+    const content = file.content ?? "";
+    return content.trim().split("\n");
+  };
+}
 
 describe("transpile", function () {
   test("ECMAScript types", () => {
@@ -174,5 +197,76 @@ describe("transpile", function () {
       'import { Foo } from "foo";',
       "export declare function foo(): Foo;",
     ]);
+  });
+
+  describe("print with tagged template literal", function () {
+    test("one line with symbol", () => {
+      const linesOf = transpile((schema) => {
+        const f = schema.generateFile("test.ts");
+        const Foo = f.import("Foo", "foo");
+        f.print`export function foo(): ${Foo} { return new ${Foo}(); };`;
+      });
+      expect(linesOf("test.ts")).toStrictEqual([
+        'import {Foo} from "foo";',
+        "",
+        "export function foo(): Foo { return new Foo(); };",
+      ]);
+    });
+    test("multi lines with symbol", () => {
+      const linesOf = transpile((schema) => {
+        const f = schema.generateFile("test.ts");
+        const Foo = f.import("Foo", "foo");
+        f.print`export function foo(): ${Foo} {
+  return new ${Foo}();
+};`;
+      });
+      expect(linesOf("test.ts")).toStrictEqual([
+        'import {Foo} from "foo";',
+        "",
+        "export function foo(): Foo {",
+        "  return new Foo();",
+        "};",
+      ]);
+    });
+    test("with empty lines", () => {
+      const linesOf = transpile((schema) => {
+        const f = schema.generateFile("test.ts");
+        const Foo = f.import("Foo", "foo");
+        f.print`
+export function foo(): ${Foo} {
+
+  return new ${Foo}();
+};
+`;
+      });
+      expect(linesOf("test.ts")).toStrictEqual([
+        'import {Foo} from "foo";',
+        "",
+        "",
+        "export function foo(): Foo {",
+        "",
+        "  return new Foo();",
+        "};",
+      ]);
+    });
+    test("empty literal", () => {
+      const linesOf = transpile((schema) => {
+        const f = schema.generateFile("test.ts");
+        f.print``;
+      });
+      expect(linesOf("test.ts")).toStrictEqual([""]);
+    });
+    test("with only symbol", () => {
+      const linesOf = transpile((schema) => {
+        const f = schema.generateFile("test.ts");
+        const Foo = f.import("Foo", "foo");
+        f.print`${Foo}`;
+      });
+      expect(linesOf("test.ts")).toStrictEqual([
+        'import {Foo} from "foo";',
+        "",
+        "Foo",
+      ]);
+    });
   });
 });
